Saving progress on Wanda_Test again
I have made some correction to the layout of some assets on the map, elevated the terrain at the right of the map to help smooth the ramp to the CarnivalZone. Added 2 textures : SM_Garland_BaseColor and SM_Garland_Alpha to make the SM_Garland material and reimported a new version of the SM_CarnivalFances. Finally, I've added another hole at the south of the edge map.